+++ py/desispec/qa/__init__.py
@@ -1,5 +1,6 @@
 from __future__ import absolute_import
 
-from .qa_exposure import QA_Frame, QA_Exposure
+from .qa_exposure import QA_Exposure
+from .qa_frame import QA_Frame
 from .qa_brick import QA_Brick
